requirements.txt
Flask
boto3
langchain
python-dotenv
faiss-cpu
requests
streamlit-lottie
uuid
langchain-community